+++ modx.ddtools.class.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * EvolutionCMS.libraries.ddTools
- * @version 0.38.1 (2020-06-02)
+ * @version 0.39 (2020-06-07)
  * 
  * @see README.md
  * 
@@ -238,11 +238,11 @@ class ddTools {
 	
 	/**
 	 * explodeAssoc
-	 * @version 1.1.5 (2019-06-22)
+	 * @version 1.1.6 (2020-06-07)
 	 * 
 	 * @desc Splits string on two separators in the associative array.
 	 * 
-	 * @param $inputString {string_separated} — String to explode. @required
+	 * @param $inputString {stringSeparated} — String to explode. @required
 	 * @param $itemDelimiter {string} — Separator between pairs of key-value. Default: '||'.
 	 * @param $keyValDelimiter {string} — Separator between key and value. Default: '::'.
 	 * 
@@ -256,7 +256,9 @@ class ddTools {
 		$result = [];
 		
 		//Если строка пустая, выкидываем сразу
-		if ($inputString == ''){return $result;}
+		if ($inputString == ''){
+			return $result;
+		}
 		
 		//Разбиваем по парам
 		$inputString = explode(
@@ -1184,13 +1186,47 @@ class ddTools {
 	}
 	
 	/**
+	 * createDocument_prepareAlias
+	 * @version 0.1 (2020-06-07)
+	 * 
+	 * @desc Translate strings.
+	 * 
+	 * @param $sourceString {string} — Document pagetitle. @required
+	 * 
+	 * @return {string} — Translated string.
+	 */
+	private static function createDocument_prepareAlias($sourceString){
+		$result = $sourceString;
+		
+		$result = transliterator_transliterate(
+			'Any-Latin; NFD; [:Nonspacing Mark:] Remove; NFC; [:Punctuation:] Remove; Lower();',
+			$result
+		);
+		
+		$result = str_replace(
+			' ',
+			'-',
+			$result
+		);
+		
+		$result = preg_replace(
+			'/[^A-Za-zА-Яа-я0-9\-_]/',
+			'',
+			$result
+		);
+		
+		return $result;
+	}
+	
+	/**
 	 * createDocument
-	 * @version 1.3.2 (2020-05-28)
+	 * @version 1.4 (2020-06-07)
 	 * 
 	 * @desc Create a new document.
 	 * 
 	 * @param $docData {stdClass|arrayAssociative} — Array of document fields or TVs. Key — name, value — value. @required
 	 * @param $docData->pagetitle {string} — Document pagetitle. Default: 'New resource'.
+	 * @param $docData->alias {string} — Document alias. If empty, will be transliterated from `$docData->pagetitle`. Default: ''.
 	 * @param $docGroups {array} — Array of document groups id.
 	 * 
 	 * @return {integer|false} — ID нового документа или false, если что-то не так.
@@ -1203,6 +1239,8 @@ class ddTools {
 		$docData = (object) array_merge(
 			[
 				'pagetitle' => 'New resource',
+				//Autotransliterate from pagetitle
+				'alias' => '',
 				//Если не передана дата создания документа, ставим текущую
 				'createdon' => time(),
 				//Если не передано, кем документ создан, ставим 1
@@ -1220,6 +1258,12 @@ class ddTools {
 		if ($docData->published == 1){
 			$docData->pub_date = $docData->createdon;
 		}
+		
+		if (trim($docData->alias) == ''){
+			$docData->alias = self::createDocument_prepareAlias($docData->pagetitle);
+		}
+		
+		$docAlias = $docData->alias;
 		
 		$docData = self::prepareDocData([
 			'data' => $docData,
@@ -1300,11 +1344,6 @@ class ddTools {
 			isset($docData->fieldsData['isfolder']) ?
 			$docData->fieldsData['isfolder'] :
 			0
-		;
-		$docAlias =
-			isset($docData->fieldsData['alias']) ?
-			$docData->fieldsData['alias'] :
-			''
 		;
 		
 		//Пусть созданного документа